Fix compiler warning.
* subversion/libsvn_fs_util/fs-util.c
  (svn_fs__canonicalize_abspath): Change variable type to size_t instead 
   of int.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@953305 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs_util/fs-util.c
@@ -39,7 +39,7 @@ svn_fs__canonicalize_abspath(const char *path, apr_pool_t *pool)
 {
   char *newpath;
   size_t path_len;
-  int path_i = 0, newpath_i = 0;
+  size_t path_i = 0, newpath_i = 0;
   svn_boolean_t eating_slashes = FALSE;
 
   /* No PATH?  No problem. */
